+++ src/utils/remove-prefix.js
@@ -1,3 +1,0 @@
-export default function removePrefix(str, prefix) {
-  return str.indexOf(prefix) === 0 ? str.slice(prefix.length) : str;
-}